c: add epoll applet.
The epoll applet shows up how to use epoll mecanism. It also could be a great
start for an epoll based application.

It displays on stdout what it can read on stdin. It is a kind of a limited
"cat" application.

$ make epoll
$ ./epoll -h
make: 'epoll' is up to date.
Usage: epoll [OPTIONS] FILE

Options:
 -t or --timeout SECOND Set time-out in seconds [default=-1].
 -D or --debug          Turn on debug messages.
 -h or --help           Display this message.
 -v or --version        Display the version.

$ ./epoll
$ cat epoll.c | ./epoll

Signed-off-by: Gaël PORTAY <gael.portay@gmail.com>