shell: add amount script
The script amount is sd device auto-mounter script. It provides a way to
run scripts afterward by adding scripts under $PREFIX/share/amount.d/
directory.

It comes with a udev rule and an sd.action script to trigger the
(un)mountage of the device.

The limitation here is that udev does not know the user that is
currently logged-in. The Makefile will give to the udev rule the
username so that the sd.action script (triggered by udev) will be able
to set the appropriate uid/gid options to pass to mount.