openssl: add client certificate authentitcation
This template shows how to set up a client certificate authentication on
a lighttpd webserver.

It uses a docker image generated by sebp [1] and it follows the tutorial
written by schnouki [2].

Test it using the command below. Enter the password to encrypt the p12
certificate file:

	$ make

Do not forget to add the root/client.p12 file to your web-browser and to
thrust the website.

[1] https://hub.docker.com/r/sebp/lighttpd/
[2] https://schnouki.net/posts/2014/08/12/lighttpd-and-ssl-client-certificates/