qemu: add runqemu.mk
This runqemu makefile provides a facility to run a linux bzImage within
Qemu system.

The kernel command-line can be specified by overriding the CMDLINE
variable through the command-line of make.

	$ make runqemu CMDLINE=debug

Because the linux kernel is not setup to have a builtin console serial
port driver, there is nothing to display on the Qemu serial console.