qemu: add initramfs.mk
The initramfs makefile creates a CPIO archive that can be used by a
linux kernel as an initial RAMFS or initial RAM Disk as rootfs.

It also makes the console node into /dev.

Because the linux kernel is not setup to run ELF binaries, Qemu still
ends with the same error.

	Kernel panic - not syncing: No working init found.  Try passing init= option to kernel. See Linux Documentation/admin-guide/init.rst for guidance.
	Kernel Offset: disabled
	---[ end Kernel panic - not syncing: No working init found.  Try passing init= option to kernel. See Linux Documentation/admin-guide/init.rst for guidance.