qemu: enable ELF support
The ELF (Executable and Linkable Format) is a format for libraries and
executables used across different architectures and operating systems.
It is the format used on Linux distributions.

The userland is started. The linux kernel has found a runnable init from
the Busybox suite.

Now Qemu reports the following errors from Busybox init.

	can't run '/etc/init.d/rcS': No such file or directory
	can't open /dev/tty3: No such file or directory
	can't open /dev/tty4: No such file or directory
	can't open /dev/tty2: No such file or directory

When there is no /etc/inittab file in the rootfs, Busybox init uses an
hardcoded one which can be summup as follow to explains these four
errors.

	...
	#
	# Boot-time system configuration/initialization script.
	# This is run first except when booting in single-user mode.
	#
	::sysinit:/etc/init.d/rcS

	# /bin/sh invocations on selected ttys
	#
	# Note below that we prefix the shell commands with a "-" to indicate to the
	# shell that it is supposed to be a login shell.  Normally this is handled by
	# login, but since we are bypassing login in this case, BusyBox lets you do
	# this yourself...
	#
	# Start an "askfirst" shell on the console (whatever that may be)
	::askfirst:-/bin/sh
	# Start an "askfirst" shell on /dev/tty2-4
	tty2::askfirst:-/bin/sh
	tty3::askfirst:-/bin/sh
	tty4::askfirst:-/bin/sh
	...

First, /etc/init.d/rcS is run; but it is missing from the initramfs.cpio
archive.

Then, three interactives /bin/sh are askfirst on tty2 to tty4; but they
complains about missing /dev/ttyx devices.