qemu: enable 8250 console support
Qemu emulates a 8250 serial port on most architectures; including Intel
x86.

Usually, those devices are named ttySx. The linux console=<DEVNAME>
parameter tells the kernel to use DEVNAME as console.

This new console can be specified to Qemu by overriding the kernel
command-line using CMDLINE variable when make is run.

	$ make runqemu CMDLINE+=console=ttyS0

Because the linux kernel is not setup to print things, it has still
nothing to display on the Qemu serial console.