qemu: add /etc/passwd /etc/group and /root
It is not possible to login because the root password is not set; and
more specifically because the root accound is not defined at all.

	(none) login: root
	Password:
	Login incorrect
	(none) login:

User accounts are defined in the /etc/passwd file and groups are defined
in /etc/group.

Let define those two files to setup both user and group for root with
/root as home directory.

It is now possible to login using tty5 or tt6 getties.

	(none) login: root
	login[25]: root login on 'tty5'
	~ # pwd
	/root